refactor(consumer-typecheck): consolidate 3 snapshot scripts into one CLI (SD-3213b)
Three standalone snapshot scripts (super-editor-package, legacy
subpaths, root 4-source inventory) shared no code but had identical
CLI boilerplate and three separate CI step blocks. Consolidate into a
single CLI with a family-module pattern.

- tests/consumer-typecheck/snapshot.mjs is the unified entry point.
  Dispatches on --all or --family <name>, with --check (default) or
  --write. Aggregates exit codes for --all.
- Each family module under tests/consumer-typecheck/snapshot/ exports
  FAMILY, DESCRIPTION, and run({ mode }) -> { code }. No shared
  abstraction; each family keeps its own enumeration and snapshot
  shape.
- ci-superdoc.yml, release-superdoc.yml, release-stable.yml collapse
  two step blocks into one snapshot.mjs --all --check call.
- CONTRIBUTING.md, snapshots/README.md, deep-type-audit.README.md,
  packages/superdoc/AGENTS.md updated to reference the unified CLI.

Verified locally: all three families pass --check against committed
snapshots, both individually and via --all.
