+++ .github/workflows/release-superdoc.yml
@@ -149,15 +149,11 @@ jobs:
         # the packed tarball. Same step as PR CI.
         run: node tests/consumer-typecheck/package-shape-gate.mjs
 
-      - name: Legacy public no-growth gates (SD-3176)
+      - name: Public surface no-growth snapshots (SD-3176, SD-3212)
         # Same gate as PR CI. Catches releases that bypass PR CI.
-        run: |
-          node tests/consumer-typecheck/snapshot-super-editor-package-exports.mjs --check
-          node tests/consumer-typecheck/snapshot-superdoc-legacy-exports.mjs --check
-
-      - name: Root no-growth + 4-source inventory (SD-3212 PR A0)
-        # Same gate as PR CI. Catches releases that bypass PR CI.
-        run: node tests/consumer-typecheck/snapshot-superdoc-root-exports.mjs --check
+        # Runs the unified entry point for all three snapshot families
+        # (super-editor-package, legacy, root).
+        run: node tests/consumer-typecheck/snapshot.mjs --all --check
 
       - name: Root classification closure gate (SD-3212 PR A1b)
         # Same gate as PR CI. Catches releases that bypass PR CI.
